refactor(lightspeed): split e2e, share suite for legacy and NFS
- Replace monolithic lightspeed.test.ts with lightspeed.ui, .mcp, and
  .conversation specs; add bootstrapLightspeedE2eSetup for shared mocks/login.
- Move e2e-tests to workspace root; Playwright testDir e2e-tests; fix translation
  and fixture paths. Remove utils/login.ts; guest flow in setup helper.
- package.json: test:e2e:nfs, test:e2e:all, playwright script runs both modes.
- Guest post-login: assert Red Hat Catalog heading only when APP_MODE is not nfs.
- expectBackstagePageVisible no-ops on nfs (no legacy catalog chrome).
- Ignore e2e-tests in ESLint for lint-staged; update knip path.

Made-with: Cursor
