+++ workspaces/lightspeed/e2e-tests/fixtures/uploads/fr.upload1.json
