+++ workspaces/lightspeed/e2e-tests/fixtures/uploads/fr.upload2.json
