+++ workspaces/lightspeed/e2e-tests/pages/McpConfigureTokenPage.ts
