+++ workspaces/lightspeed/e2e-tests/utils/accessibility.ts
