+++ workspaces/lightspeed/e2e-tests/utils/chatManagement.ts
