+++ workspaces/lightspeed/e2e-tests/utils/fileUpload.ts
