+++ workspaces/lightspeed/e2e-tests/utils/sidebar.ts
