+++ workspaces/lightspeed/package.json
@@ -19,11 +19,12 @@
     "clean": "backstage-cli repo clean",
     "test": "backstage-cli repo test",
     "test:all": "backstage-cli repo test --coverage",
-    "test:e2e": "yarn test:e2e:legacy",
+    "test:e2e": "yarn test:e2e:all",
     "test:e2e:legacy": "APP_MODE=legacy playwright test",
-    "test:e2e:all": "yarn test:e2e:legacy",
+    "test:e2e:nfs": "APP_MODE=nfs playwright test",
+    "test:e2e:all": "yarn test:e2e:legacy && yarn test:e2e:nfs",
     "test:e2e:ci": "yarn test:e2e:all",
-    "playwright": "bash -c 'if [[ $@ == test ]]; then yarn test:e2e:legacy; else npx playwright \"$@\"; fi' _",
+    "playwright": "bash -c 'if [[ $@ == test ]]; then yarn test:e2e:all; else npx playwright \"$@\"; fi' _",
     "fix": "backstage-cli repo fix",
     "lint": "backstage-cli repo lint --since origin/main",
     "lint:all": "backstage-cli repo lint",
